+++ src/react/components/banner/XUIBanner.js
@@ -5,12 +5,14 @@ import crossIcon from '@xero/xui-icon/icons/cross';
 import XUIIconButton from '../button/XUIIconButton';
 import sentimentMap from './private/sentiments';
 import { ns } from '../helpers/xuiClassNamespace';
+import checkRequiredProps from '../../helpers/checkRequiredProps';
 
 const sentiments = Object.keys(sentimentMap);
 
 const XUIBanner = ({
   children,
   className,
+  closeButtonLabel,
   hasDefaultLayout,
   onCloseClick,
   qaHook,
@@ -19,13 +21,13 @@ const XUIBanner = ({
 }) => {
   const closeButton = onCloseClick && (
     <XUIIconButton
-      ariaLabel="Close"
+      ariaLabel={closeButtonLabel}
       className={`${ns}-banner--close`}
       icon={crossIcon}
       onClick={onCloseClick}
       qaHook={qaHook && `${qaHook}-close--button`}
       size="small"
-      title="Close"
+      title={closeButtonLabel}
     />
   );
 
@@ -55,6 +57,15 @@ export default XUIBanner;
 XUIBanner.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
+  /**
+   * Title and accessibility label to be applied to the banner close "X" button.
+   * This is required if an `onCloseClick` callback prop is provided.
+   * <br />
+   * Recommended English value: *Close*
+   */
+  closeButtonLabel(...parameters) {
+    return checkRequiredProps('onCloseClick', PropTypes.string.isRequired, ...parameters);
+  },
   /** Defines whether the default layout class should be supplied */
   hasDefaultLayout: PropTypes.bool,
   /** Handles the click event for the action */