+++ src/react/components/banner/stories/stories.js
@@ -53,6 +53,7 @@ storiesWithKnobs.add('Playground', () => (
       },
     ])}
     className={text('className', '')}
+    closeButtonLabel="Close"
     detailItems={array('detailItems', [])}
     hasDefaultLayout={boolean('hasLayout', true)}
     messageText={text('messageText', "A Banner's Message")}
@@ -69,6 +70,6 @@ variations.forEach(variation => {
     variationMinusStoryDetails.storyKind = undefined;
     variationMinusStoryDetails.storyTitle = undefined;
 
-    return <DetailedBanner {...variationMinusStoryDetails} />;
+    return <DetailedBanner closeButtonLabel="Close" {...variationMinusStoryDetails} />;
   });
 });