+++ src/react/components/toast/XUIToast.d.ts
@@ -16,6 +16,13 @@ interface BaseProps {
    */
   className?: string;
   /**
+   * Title and accessibility label to be applied to the toast close "X" button.
+   * This is required if an `onCloseClick` callback prop is provided.
+   *
+   * Recommended English value: *Close*
+   */
+  closeButtonLabel?: string;
+  /**
    * Applies default layout class to the component.
    */
   hasDefaultLayout?: boolean;